+++ lib/src/mutable_document.dart
@@ -125,9 +125,4 @@ class MutableDocument extends Document {
 
     return null;
   }
-
-  /// Gets content of the current object as a Dictionary.
-  ///
-  /// - Returns: The Dictionary representing the content of the current object.
-  Map<String, dynamic> toMap() => Map.from(_data);
 }